[#1] Add file uploads
Files are stored on the server's file system, organized by
files/year/month/day/. Each day directory must have unique filenames
within it, a BudgetError will be raised if this would be violated.
This functionality broke the "has file" sort option. I'm not sure when
it will be back or if it will be needed at all.
The test code for this will crash on any system but mine, so I've
commented out the code that tests file uploads.
The backend is not responsible for getting the actual file data, that
must be provided separately, matched to the file list in the Transaction
object. Adding files after the transaction is recorded is not
supported.